follow, add people

+++ follow.php
@@ -25,51 +25,49 @@ session_start();
     //echo 'Fodllowing ID: '. $followingID .'<br>';
     //echo 'Fodllowing Name: '. $followingName .'<br>';
 
+    if($followerID !=="" && $followingID !=="" && $followingName !=="" && $followerName !=="") {
+
+        //check if the user is already following
+        $check = mysqli_query($connect, "SELECT * FROM `followingx` WHERE follower_id = '$followingID' AND `user_id`='$followerID'");
+        $countCheck = mysqli_num_rows($check);
+
+        if ($countCheck <= 0) {
+
+            $insert = @mysqli_query($connect, "INSERT INTO `followingx` (`username`,`user_id`,`follower_name`,`follower_id`,`status`,`date`) VALUES('" . $followerName . "','" . $followerID . "','" . $followingName . "','" . $followingID . "','0','" . $date . "')");
 
 
-    //check if the user is already following
+            //notify the person of the follower information
 
-    $check = mysqli_query($connect,"SELECT * FROM `followingx` WHERE follower_id = '$followingID' AND `user_id`='$followerID'");
-
-    $countCheck = mysqli_num_rows($check);
-
-    if($countCheck <= 0){
-
-        $insert = @mysqli_query($connect,"INSERT INTO `followingx` (`username`,`user_id`,`follower_name`,`follower_id`,`status`,`date`) VALUES('".$followerName."','". $followerID ."','". $followingName."','". $followingID ."','0','". $date .  "')");
+            $message = $followerName . 'Started following you.. check <a href="profile/check/' . $followerID . '">' . $followingName . 's profile for more information</a>';
+            $link = "profile/check/" . $followingID;
+            $user_id = $followerID;
+            $status = 0;
+            $remake = '';
 
 
-        //notify the person of the follower information
-
-        $message = $followerName. 'Started following you.. check <a href="profile/check/'. $followerID .'">'. $followingName .'s profile for more information</a>';
-        $link = "profile/check/".$followingID;
-        $user_id = $followerID;
-        $status = 0;
-        $remake = '';
+            $insertNotification = mysqli_query($connect, "INSERT INTO `notificationx` (`message`,`link`,`user_id`,`status`,`date`,`remark`) VALUES('$message','" . $link . "','$user_id','0','$date','$remake')");
 
 
-        $insertNotification = mysqli_query($connect, "INSERT INTO `notificationx` (`message`,`link`,`user_id`,`status`,`date`,`remark`) VALUES('$message','".$link."','$user_id','0','$date','$remake')");
+            if (!$insertNotification) {
 
+                echo mysqli_error($connect);
 
-        if(!$insertNotification){
+            }
 
-            echo mysqli_error($connect);
-
+            echo '<a class="btn follow following" href="#"> <i class="ext-red"></i>Following</a>';
         }
+        else {
 
-        echo '<a class="btn follow following" href="#"> <i class="ext-red"></i>Following</a>';
-    }
+            $deleteFolower = mysqli_query($connect, "DELETE FROM `followingx` WHERE follower_id = '$followingID' AND `user_id`='$followerID'");
 
-    else{
+            if (!$deleteFolower) {
+                echo mysqli_error($connect);
+            }
 
-        $deleteFolower = mysqli_query($connect,"DELETE FROM `followingx` WHERE follower_id = '$followingID' AND `user_id`='$followerID'");
+            echo '<a class="btn follow" href="#"> <i class="fa fa-user-plus text-red"></i> Follow</a>';
 
-        if(!$deleteFolower){
-            echo mysqli_error($connect);
+            //die('Delete');
         }
-
-        echo '<a class="btn follow" href="#"> <i class="fa fa-user-plus text-red"></i> Follow</a>';
-
-        //die('Delete');
     }
